Update validation method
Validate integer seems to not like 01 - I've tried quoting and not quoting and it makes no difference.

Update dependency on auditd package

With dependency set as Concat['audit-file'], the module does not work. Setting to package but could be anything else appropriate.

==> proxy01: Error: Cannot create /etc/audit/rules.d; parent directory /etc/audit does not exist
==> proxy01: Error: /Stage[main]/Auditd/File[/etc/audit/rules.d]/ensure: change from absent to directory failed: Cannot create /etc/audit/rules.d; parent directory /etc/audit does not exist

Add operating system release for spec tests to pass.