Add support for audisp-cef plugin
`audisp-cef` is an audisp plugin which aggregates audit messages with the same
ID into a single syslog message of the Common Event Format.

The plugin is published on Github and the makefile provides targets for making
.deb and .npm packages. At this time these packages do not appear to be
provided by the default repositories in Ubuntu or CentOS. Users wishing to
deploy this plugin can in these cases provide the `audisp-cef` package
with their own repository.

Source code for the plugin: https://github.com/gdestuynder/audisp-cef